Updated Project to get Android device testing to work
There are a few settings I needed to change to test via Android device. You can now do this by downloading the app Unity Remote, plugging into the computer via USB and enabling USB debugging. If you then click run and build you can test on your phone. I decided to do this as it's faster then generating an apk each time and downloading it on the phone. Not sure how this works with apple phones but we can look at that later if we need to.